requirements.txt
discord.py
python-dotenv==0.12.0